Replace old metrics with the new ones
