+++ seismiqb/src/metrics.py
@@ -1,63 +1,63 @@
-""" Contains metrics for various labels (horizons, facies, etc) and cubes. """
-#pylint: disable=too-many-lines, not-an-iterable
+""" Metrics for seismic objects: cubes and horizons. """
 from copy import copy
 from textwrap import dedent
 from tqdm.auto import tqdm
 
 import numpy as np
-from numba import njit, prange
-import matplotlib.colors as mcolors
+try:
+    import cupy as cp
+except ImportError:
+    cp = np
 
 import cv2
-from scipy.signal import hilbert, medfilt
-from scipy.stats import mode as mode_scipy
+from matplotlib.colors import LinearSegmentedColormap
+from matplotlib.cm import register_cmap
 
-from ..batchflow.models.metrics import Metrics
+from ..batchflow.notifier import Notifier
 
 from .horizon import Horizon
-from .utils import mode, compute_running_mean
+from .utility_classes import Accumulator
+from .functional import to_device, from_device
+from .functional import correlation, crosscorrelation, btch, kl, js, hellinger, tv, hilbert
+from .functional import smooth_out, digitize, gridify, perturb, histo_reduce
 from .plotters import plot_image
 
 
 
 CDICT = {
-    'red': [
-        [0.0, None, 1.0],
-        [0.33, 1.0, 1.0],
-        [0.66, 1.0, 1.0],
-        [1.0, 0.0, None]
-    ],
-    'green': [
-        [0.0, None, 0.0],
-        [0.33, 0.0, 0.0],
-        [0.66, 1.0, 1.0],
-        [1.0, 0.5019607843137255, None]
-    ],
-    'blue': [
-        [0.0, None, 0.0],
-        [0.33, 0.0, 0.0],
-        [0.66, 0.0, 0.0],
-        [1.0, 0.0, None]
-    ]
+    'red': [[0.0, None, 1.0], [0.33, 1.0, 1.0], [0.66, 1.0, 1.0], [1.0, 0.0, None]],
+    'green': [[0.0, None, 0.0], [0.33, 0.0, 0.0], [0.66, 1.0, 1.0], [1.0, 0.5, None]],
+    'blue': [[0.0, None, 0.0], [0.33, 0.0, 0.0], [0.66, 0.0, 0.0], [1.0, 0.0, None]]
 }
-METRIC_CMAP = mcolors.LinearSegmentedColormap('CustomMap', CDICT)
+METRIC_CMAP = LinearSegmentedColormap('MetricMap', CDICT)
+METRIC_CMAP.set_bad(color='black')
+register_cmap(name='Metric', cmap=METRIC_CMAP)
 
 
 
-
-class BaseSeismicMetric(Metrics):
+class BaseMetrics:
     """ Base class for seismic metrics.
-    Child classes have to implement access to `data`, `probs`, `spatial` and `cube_name` attributes.
+    Child classes have to implement access to `data`, `probs`, `bad_traces` attributes.
     """
+    # pylint: disable=attribute-defined-outside-init, blacklisted-name
+    PLOT_DEFAULTS = {
+        'cmap': 'Metric',
+        'fill_color': 'black'
+    }
+
     LOCAL_DEFAULTS = {
         'kernel_size': 3,
-        'reduce_func': 'nanmean',
-        'agg': None,
+        'agg': 'nanmean',
+        'device': 'gpu',
+        'amortize': True,
     }
 
     SUPPORT_DEFAULTS = {
-        'supports': 20,
-        'agg': 'mean',
+        'supports': 100,
+        'safe_strip': 50,
+        'agg': 'nanmean',
+        'device': 'gpu',
+        'amortize': True,
     }
 
     SMOOTHING_DEFAULTS = {
@@ -67,417 +67,489 @@ class BaseSeismicMetric(Metrics):
 
     EPS = 0.00001
 
-    def evaluate(self, metric, agg=None, plot=False, show_plot=True, savepath=None, backend='matplotlib', **kwargs):
+
+    def evaluate(self, metric, plot=False, plot_supports=False, enlarge=True, width=5, **kwargs):
         """ Calculate desired metric, apply aggregation, then plot resulting metric-map.
         To plot the results, set `plot` argument to True.
 
         Parameters
         ----------
-        metric : str or sequence of str
+        metric : str
             Name of metric to evaluate.
-        agg : int, str or callable
-            Function to transform metric from ndarray of (n_ilines, n_xlines, N) shape to (n_ilines, n_xlines) shape.
-            If callable, then directly applied to the output of metric computation function.
-            If str, then must be a function from `numpy` module. Applied along the last axis only.
-            If int, then index of slice along the last axis to return.
-        plot, show_plot, savepath, backend, plot_kwargs
-            Parameters that are passed directly to plotting function, see :func:`.plot_image`.
+        enlarge : bool
+            Whether to apply `:meth:.Horizon.enlarge_carcass_image` to the result.
+        width : int
+            Widening for the metric. Works only if `enlarge` set to True.
+        plot : bool
+            Whether to use `:func:.plot_image` to show the result.
+        plot_supports : bool
+            Whether to show support traces on resulting image. Works only if `plot` set to True.
         kwargs : dict
-            Metric-specific parameters.
+            Arguments to be passed in metric-calculation methods
+            (see `:meth:.compute_local` and `:meth:.compute_support`),
+            as well as plotting arguments (see `:func:.plot_image`).
         """
-        if agg is None:
-            if 'support' in metric:
-                agg = 'nanmean'
+        if 'support' in metric:
+            kwargs = {**self.SUPPORT_DEFAULTS, **kwargs}
+        elif 'local' in metric:
+            kwargs = {**self.LOCAL_DEFAULTS, **kwargs}
 
-        # Get metric, then aggregate
+        self._last_evaluation = {}
         metric_fn = getattr(self, metric)
         metric_val, plot_dict = metric_fn(**kwargs)
-        metric_val = self._aggregate(metric_val, agg)
 
-        # Get plot parameters
+        if cp is not np:
+            # pylint: disable=protected-access
+            cp._default_memory_pool.free_all_blocks()
+
+        if hasattr(self, 'horizon') and self.horizon.is_carcass and enlarge:
+            metric_val = self.horizon.enlarge_carcass_image(metric_val, width)
+
         if plot:
-            spatial = plot_dict.pop('spatial', True)
-            ignore_value = plot_dict.pop('ignore_value', None)
+            plot_dict = {**self.PLOT_DEFAULTS, **plot_dict}
+            figure = plot_image(metric_val, **plot_dict, return_figure=True)
 
-            # np.nan allows to ignore values
-            if ignore_value is not None:
-                copy_metric = np.copy(metric_val)
-                copy_metric[copy_metric == ignore_value] = np.nan
-            else:
-                copy_metric = metric_val
+            if 'support' in metric and plot_supports:
+                support_coords = self._last_evaluation['support_coords']
+                figure.axes[0].scatter(support_coords[:, 0],
+                                       support_coords[:, 1], s=33, marker='.', c='blue')
 
-        # Actual plot
-            if spatial:
-                plot_image(copy_metric, savepath=savepath, show=show_plot, backend=backend, **plot_dict)
-            else:
-                pass
+            # Store for debug / introspection purposes
+            self._last_evaluation['plot_dict'] = plot_dict
+            self._last_evaluation['figure'] = figure
         return metric_val
 
-    def _aggregate(self, metric, agg=None):
-        if agg is not None:
-            if callable(agg):
-                metric = agg(metric)
-            elif isinstance(agg, str):
-                if agg == 'mode':
-                    metric = mode(metric)
-                else:
-                    metric = getattr(np, agg)(metric, axis=-1)
-            elif isinstance(agg, (int, slice)):
-                metric = metric[..., agg]
-        return metric
 
+    def compute_local(self, function, data, bad_traces, kernel_size=3,
+                      normalize=True, agg='mean', amortize=False, axis=0, device='cpu', pbar=None):
+        """ Compute metric in a local fashion, using `function` to compare nearest traces.
+        Under the hood, each trace is compared against its nearest neighbours in a square window
+        of `kernel_size` size. Results of comparisons are aggregated via `agg` function.
 
-    def local_corrs(self, kernel_size=3, reduce_func='nanmean', **kwargs):
-        """ Compute correlation between each column in data and nearest traces.
+        Works on both `cpu` (via standard `NumPy`) and GPU (with the help of `cupy` library).
+        The returned array is always on CPU.
 
         Parameters
         ----------
+        function : callable
+            Function to compare two arrays. Must have the following signature:
+            `(array1, array2, std1, std2)`, where `std1` and `std2` are pre-computed standard deviations.
+            In order to work properly on GPU, must be device-agnostic.
+        data : ndarray
+            3D array of data to evaluate on.
+        bad_traces : ndarray
+            2D matrix of traces where the metric should not be computed.
         kernel_size : int
-            Size of window to reduce values in.
-        reduce_func : str or callable
-            Function to reduce values in window with, e.g. `mean` or `nanmax`.
-
-        Returns
-        -------
-        array-like
-            Matrix of (n_ilines, n_xlines) shape with computed metric for each point.
+            Window size for comparison traces.
+        normalize : bool
+            Whether the data should be zero-meaned before computing metric.
+        agg : str
+            Function to aggregate values for each trace. See :class:`.Accumulator` for details.
+        amortize : bool
+            Whether the aggregation should be sequential or by stacking all the matrices.
+            See :class:`.Accumulator` for details.
+        axis : int
+            Axis to stack arrays on. See :class:`.Accumulator` for details.
+        device : str
+            Device specificator. Can be either string (`cpu`, `gpu:4`) or integer (`4`).
+        pbar : type or None
+            Progress bar to use.
         """
-        metric, title = compute_local_corrs(data=self.data, bad_traces=self.bad_traces,
-                                            kernel_size=kernel_size, reduce_func=reduce_func, **kwargs)
-        plot_dict = {
-            'spatial': self.spatial,
-            'title': f'{title} for {self.name} on cube {self.cube_name}, k={kernel_size}, reduce={reduce_func}',
-            'cmap': METRIC_CMAP,
-            'zmin': -1, 'zmax': 1,
-            'ignore_value': 0.0,
-            'xlabel': 'INLINE_3D', 'ylabel': 'CROSSLINE_3D',
-            'fill_color': 'black',
-            **kwargs
-        }
-        return metric, plot_dict
+        i_range, x_range = data.shape[:2]
+        k = kernel_size // 2
 
-    def support_corrs(self, supports=10, safe_strip=0, line_no=None, **kwargs):
-        """ Compute correlations between each trace and support traces.
+        # Transfer to GPU, if needed
+        data = to_device(data, device)
+        bad_traces = to_device(bad_traces, device)
+        xp = cp.get_array_module(data) if (cp is not np) else np
+
+        # Compute data statistics
+        data_stds = data.std(axis=-1)
+        bad_traces[data_stds == 0.0] = 1
+        if normalize:
+            data_n = data - data.mean(axis=-1, keepdims=True)
+        else:
+            data_n = data
+
+        # Pad everything
+        padded_data = xp.pad(data_n, ((k, k), (k, k), (0, 0)), constant_values=xp.nan)
+        padded_stds = xp.pad(data_stds, k, constant_values=0.0)
+        padded_bad_traces = xp.pad(bad_traces, k, constant_values=1)
+
+        # Compute metric by shifting arrays
+        total = kernel_size * kernel_size - 1
+        pbar = Notifier(pbar, total=total) if pbar else None
+
+        accumulator = Accumulator(agg=agg, amortize=amortize, axis=axis, total=total)
+        for i in range(kernel_size):
+            for j in range(kernel_size):
+                if i == j == k:
+                    continue
+
+                shifted_data = padded_data[i:i+i_range, j:j+x_range]
+                shifted_stds = padded_stds[i:i+i_range, j:j+x_range]
+                shifted_bad_traces = padded_bad_traces[i:i+i_range, j:j+x_range]
+
+                computed = function(data_n, shifted_data, data_stds, shifted_stds)
+                computed[shifted_bad_traces == 1] = xp.nan
+                accumulator.update(computed)
+                if pbar:
+                    pbar.update()
+        if pbar:
+            pbar.close()
+
+        result = accumulator.get(final=True)
+        return from_device(result)
+
+    def compute_support(self, function, data, bad_traces, supports, safe_strip=0,
+                        normalize=True, agg='mean', amortize=False, axis=0, device='cpu', pbar=None):
+        """ Compute metric in a support fashion, using `function` to compare all the traces
+        against a set of (randomly chosen or supplied) reference ones.
+        Results of comparisons are aggregated via `agg` function.
+
+        Works on both `cpu` (via standard `NumPy`) and GPU (with the help of `cupy` library).
+        The returned array is always on CPU.
 
         Parameters
         ----------
-        supports : int, sequence, ndarray or str
-            Defines mode of generating support traces.
-            If int, then that number of random non-zero traces positions are generated.
-            If sequence or ndarray, then must be of shape (N, 2) and is used as positions of support traces.
-            If str, then must define either `iline` or `xline` mode. In each respective one, iline/xline given by
-            `line_no` argument is used to generate supports.
-        safe_strip : int
-            Used only for `int` mode of `supports` parameter and defines minimum distance
-            from borders for sampled points.
-        line_no : int
-            Used only for `str` mode of `supports` parameter to define exact iline/xline to use.
-
-        Returns
-        -------
-        array-like
-            Matrix of either (n_ilines, n_xlines, n_supports) or (n_ilines, n_xlines) shape with
-            computed metric for each point.
+        function : callable
+            Function to compare two arrays. Must have the following signature:
+            `(array1, array2, std1, std2)`, where `std1` and `std2` are pre-computed standard deviations.
+            In order to work properly on GPU, must be device-agnostic.
+        data : ndarray
+            3D array of data to evaluate on.
+        bad_traces : ndarray
+            2D matrix of traces where the metric should not be computed.
+        supports : int or ndarray
+            If int, then number of supports to generate randomly from non-bad traces.
+            If ndarray, then should be of (N, 2) shape and contain coordinates of reference traces.
+        normalize : bool
+            Whether the data should be zero-meaned before computing metric.
+        agg : str
+            Function to aggregate values for each trace. See :class:`.Accumulator` for details.
+        amortize : bool
+            Whether the aggregation should be sequential or by stacking all the matrices.
+            See :class:`.Accumulator` for details.
+        axis : int
+            Axis to stack arrays on. See :class:`.Accumulator` for details.
+        device : str
+            Device specificator. Can be either string (`cpu`, `gpu:4`) or integer (`4`).
+        pbar : type or None
+            Progress bar to use.
         """
-        metric, title = compute_support_corrs(data=self.data, supports=supports, bad_traces=self.bad_traces,
-                                              safe_strip=safe_strip, line_no=line_no, **kwargs)
-        plot_dict = {
-            'spatial': self.spatial,
-            'title': f'{title} for {self.name} on cube {self.cube_name}',
-            'cmap': METRIC_CMAP,
-            'zmin': -1.0, 'zmax': 1.0,
-            'ignore_value': 0.0,
-            'xlabel': 'INLINE_3D', 'ylabel': 'CROSSLINE_3D',
-            'fill_color': 'black',
-            **kwargs
-        }
-        return metric, plot_dict
+        # Transfer to GPU, if needed
+        data = to_device(data, device)
+        bad_traces = to_device(bad_traces, device)
+        xp = cp.get_array_module(data) if (cp is not np) else np
 
-
-    def local_crosscorrs(self, kernel_size=3, reduce_func='nanmean', **kwargs):
-        """ Compute cross-correlation between each column in data and nearest traces. """
-        metric, title = compute_local_crosscorrs(data=self.data, bad_traces=self.bad_traces,
-                                                 kernel_size=kernel_size, reduce_func=reduce_func, **kwargs)
-        plot_dict = {
-            'spatial': self.spatial,
-            'title': f'{title} for {self.name} on cube {self.cube_name}, k={kernel_size}, reduce={reduce_func}',
-            'cmap': 'seismic_r',
-            'zmin': None, 'zmax': None,
-            'ignore_value': np.nan,
-            'xlabel': 'INLINE_3D', 'ylabel': 'CROSSLINE_3D',
-            'fill_color': 'black',
-            **kwargs
-        }
-        return metric, plot_dict
-
-    def support_crosscorrs(self, supports=10, safe_strip=0, **kwargs):
-        """ Compute cross-correlation between each trace and support traces. """
-        metric, title = compute_support_crosscorrs(data=self.data, supports=supports, bad_traces=self.bad_traces,
-                                                   safe_strip=safe_strip, **kwargs)
-        plot_dict = {
-            'spatial': self.spatial,
-            'title': f'{title} for {self.name} on cube {self.cube_name}',
-            'cmap': 'seismic_r',
-            'zmin': None, 'zmax': None,
-            'ignore_value': np.nan,
-            'xlabel': 'INLINE_3D', 'ylabel': 'CROSSLINE_3D',
-            'fill_color': 'black',
-            **kwargs
-        }
-        return metric, plot_dict
-
-
-    def local_btch(self, kernel_size=3, reduce_func='nanmean', **kwargs):
-        """ Compute Bhattacharyya distance between each column in data and nearest traces. """
-        metric, title = compute_local_btch(data=self.probs, bad_traces=self.bad_traces,
-                                           kernel_size=kernel_size, reduce_func=reduce_func, **kwargs)
-        plot_dict = {
-            'spatial': self.spatial,
-            'title': f'{title} for {self.name} on cube {self.cube_name}, k={kernel_size}, reduce={reduce_func}',
-            'cmap': METRIC_CMAP,
-            'zmin': 0.0, 'zmax': 1.0,
-            'ignore_value': np.nan,
-            'xlabel': 'INLINE_3D', 'ylabel': 'CROSSLINE_3D',
-            'fill_color': 'black',
-            **kwargs
-        }
-        return metric, plot_dict
-
-    def support_btch(self, supports=10, safe_strip=0, **kwargs):
-        """ Compute Bhattacharyya distance between each trace and support traces. """
-        metric, title = compute_support_btch(data=self.probs, supports=supports, bad_traces=self.bad_traces,
-                                             safe_strip=safe_strip, **kwargs)
-        plot_dict = {
-            'spatial': self.spatial,
-            'title': f'{title} for {self.name} on cube {self.cube_name}',
-            'cmap': METRIC_CMAP,
-            'zmin': 0.0, 'zmax': 1.0,
-            'ignore_value': np.nan,
-            'xlabel': 'INLINE_3D', 'ylabel': 'CROSSLINE_3D',
-            'fill_color': 'black',
-            **kwargs
-        }
-        return metric, plot_dict
-
-    # Aliases for Bhattacharyya distance
-    local_bt = local_btch
-    support_bt = support_btch
-
-
-    def local_kl(self, kernel_size=3, reduce_func='nanmean', **kwargs):
-        """ Compute Kullback-Leibler divergence between each column in data and nearest traces. """
-        metric, title = compute_local_kl(data=self.probs, bad_traces=self.bad_traces,
-                                         kernel_size=kernel_size, reduce_func=reduce_func, **kwargs)
-        plot_dict = {
-            'spatial': self.spatial,
-            'title': f'{title} for {self.name} on cube {self.cube_name}, k={kernel_size}, reduce={reduce_func}',
-            'cmap': METRIC_CMAP,
-            'zmin': None, 'zmax': None,
-            'ignore_value': np.nan,
-            'xlabel': 'INLINE_3D', 'ylabel': 'CROSSLINE_3D',
-            'fill_color': 'black',
-            **kwargs
-        }
-        return metric, plot_dict
-
-    def support_kl(self, supports=10, safe_strip=0, **kwargs):
-        """ Compute Kullback-Leibler divergence between each trace and support traces. """
-        metric, title = compute_support_kl(data=self.probs, supports=supports, bad_traces=self.bad_traces,
-                                           safe_strip=safe_strip, **kwargs)
-        plot_dict = {
-            'spatial': self.spatial,
-            'title': f'{title} for {self.name} on cube {self.cube_name}',
-            'cmap': METRIC_CMAP,
-            'zmin': None, 'zmax': None,
-            'ignore_value': np.nan,
-            'xlabel': 'INLINE_3D', 'ylabel': 'CROSSLINE_3D',
-            'fill_color': 'black',
-            **kwargs
-        }
-        return metric, plot_dict
-
-
-    def local_js(self, kernel_size=3, reduce_func='nanmean', **kwargs):
-        """ Compute Jensen-Shannon distance between each column in data and nearest traces.
-        Janson-Shannon distance is a symmetrized version of Kullback-Leibler divergence.
-        """
-        metric, title = compute_local_js(data=self.probs, bad_traces=self.bad_traces,
-                                         kernel_size=kernel_size, reduce_func=reduce_func, **kwargs)
-        plot_dict = {
-            'spatial': self.spatial,
-            'title': f'{title} for {self.name} on cube {self.cube_name}, k={kernel_size}, reduce={reduce_func}',
-            'cmap': METRIC_CMAP,
-            'zmin': None, 'zmax': None,
-            'ignore_value': np.nan,
-            'xlabel': 'INLINE_3D', 'ylabel': 'CROSSLINE_3D',
-            'fill_color': 'black',
-            **kwargs
-        }
-        return metric, plot_dict
-
-    def support_js(self, supports=10, safe_strip=0, **kwargs):
-        """ Compute Jensen-Shannon distance between each trace and support traces. """
-        metric, title = compute_support_js(data=self.probs, supports=supports, bad_traces=self.bad_traces,
-                                           safe_strip=safe_strip, **kwargs)
-        plot_dict = {
-            'spatial': self.spatial,
-            'title': f'{title} for {self.name} on cube {self.cube_name}',
-            'cmap': METRIC_CMAP,
-            'zmin': None, 'zmax': None,
-            'ignore_value': np.nan,
-            'xlabel': 'INLINE_3D', 'ylabel': 'CROSSLINE_3D',
-            'fill_color': 'black',
-            **kwargs
-        }
-        return metric, plot_dict
-
-
-    def local_hellinger(self, kernel_size=3, reduce_func='nanmean', **kwargs):
-        """ Compute Hellinger distance between each column in data and nearest traces. """
-        metric, title = compute_local_hellinger(data=self.probs, bad_traces=self.bad_traces,
-                                                kernel_size=kernel_size, reduce_func=reduce_func, **kwargs)
-        plot_dict = {
-            'spatial': self.spatial,
-            'title': f'{title} for {self.name} on cube {self.cube_name}, k={kernel_size}, reduce={reduce_func}',
-            'cmap': METRIC_CMAP,
-            'zmin': None, 'zmax': None,
-            'ignore_value': np.nan,
-            'xlabel': 'INLINE_3D', 'ylabel': 'CROSSLINE_3D',
-            'fill_color': 'black',
-            **kwargs
-        }
-        return metric, plot_dict
-
-    def support_hellinger(self, supports=10, safe_strip=0, **kwargs):
-        """ Compute Hellinger distance between each trace and support traces. """
-        metric, title = compute_support_hellinger(data=self.probs, supports=supports, bad_traces=self.bad_traces,
-                                                  safe_strip=safe_strip, **kwargs)
-        plot_dict = {
-            'spatial': self.spatial,
-            'title': f'{title} for {self.name} on cube {self.cube_name}',
-            'cmap': METRIC_CMAP,
-            'zmin': None, 'zmax': None,
-            'ignore_value': np.nan,
-            'xlabel': 'INLINE_3D', 'ylabel': 'CROSSLINE_3D',
-            'fill_color': 'black',
-            **kwargs
-        }
-        return metric, plot_dict
-
-
-    def local_tv(self, kernel_size=3, reduce_func='nanmean', **kwargs):
-        """ Compute total variation distance between each column in data and nearest traces. """
-        metric, title = compute_local_tv(data=self.probs, bad_traces=self.bad_traces,
-                                         kernel_size=kernel_size, reduce_func=reduce_func, **kwargs)
-        plot_dict = {
-            'spatial': self.spatial,
-            'title': f'{title} for {self.name} on cube {self.cube_name}, k={kernel_size}, reduce={reduce_func}',
-            'cmap': METRIC_CMAP,
-            'zmin': None, 'zmax': None,
-            'ignore_value': np.nan,
-            'xlabel': 'INLINE_3D', 'ylabel': 'CROSSLINE_3D',
-            'fill_color': 'black',
-            **kwargs
-        }
-        return metric, plot_dict
-
-    def support_tv(self, supports=10, safe_strip=0, **kwargs):
-        """ Compute total variation distance between each trace and support traces. """
-        metric, title = compute_support_tv(data=self.probs, supports=supports, bad_traces=self.bad_traces,
-                                           safe_strip=safe_strip, **kwargs)
-        plot_dict = {
-            'spatial': self.spatial,
-            'title': f'{title} for {self.name} on cube {self.cube_name}',
-            'cmap': METRIC_CMAP,
-            'zmin': None, 'zmax': None,
-            'ignore_value': np.nan,
-            'xlabel': 'INLINE_3D', 'ylabel': 'CROSSLINE_3D',
-            'fill_color': 'black',
-            **kwargs
-        }
-        return metric, plot_dict
-
-
-    def local_wasserstein(self, kernel_size=3, reduce_func='nanmean', **kwargs):
-        """ Compute Wasserstein distance between each column in data and nearest traces.
-        Wasserstein distance is also known as Earth-Mover distance.
-        """
-        metric, title = compute_local_wasserstein(data=self.probs, bad_traces=self.bad_traces,
-                                                  kernel_size=kernel_size, reduce_func=reduce_func, **kwargs)
-        plot_dict = {
-            'spatial': self.spatial,
-            'title': f'{title} for {self.name} on cube {self.cube_name}, k={kernel_size}, reduce={reduce_func}',
-            'cmap': METRIC_CMAP,
-            'zmin': None, 'zmax': None,
-            'ignore_value': np.nan,
-            'xlabel': 'INLINE_3D', 'ylabel': 'CROSSLINE_3D',
-            'fill_color': 'black',
-            **kwargs
-        }
-        return metric, plot_dict
-
-    def support_wasserstein(self, supports=10, safe_strip=0, **kwargs):
-        """ Compute Wasserstein distance between each trace and support traces. """
-        metric, title = compute_support_wasserstein(data=self.probs, supports=supports, bad_traces=self.bad_traces,
-                                                    safe_strip=safe_strip, **kwargs)
-        plot_dict = {
-            'spatial': self.spatial,
-            'title': f'{title} for {self.name} on cube {self.cube_name}',
-            'cmap': METRIC_CMAP,
-            'zmin': None, 'zmax': None,
-            'ignore_value': np.nan,
-            'xlabel': 'INLINE_3D', 'ylabel': 'CROSSLINE_3D',
-            'fill_color': 'black',
-            **kwargs
-        }
-        return metric, plot_dict
-
-    # Aliases for Wasserstein distance
-    local_emd = local_wasserstein
-    support_emd = support_wasserstein
-
-
-    def hilbert(self, correction='median', kernel_size=3, eps=1e-5, **kwargs):
-        """ Compute phase along the data. """
-        _ = kwargs
-        # full_matrix = self.horizon.full_matrix
-
-        analytic = hilbert(self.data, axis=-1)
-        phase = (np.angle(analytic))
-        phase = phase % (2 * np.pi) - np.pi
-        # phase[full_matrix == Horizon.FILL_VALUE, :] = 0
-
-        horizon_phase = phase[:, :, phase.shape[-1] // 2]
-        horizon_phase = correct_pi(horizon_phase, eps)
-
-        if correction == 'mean':
-            median_phase = compute_running_mean(horizon_phase, kernel_size)
+        # Compute data statistics
+        data_stds = data.std(axis=-1)
+        bad_traces[data_stds == 0.0] = 1
+        if normalize:
+            data_n = data - data.mean(axis=-1, keepdims=True)
         else:
-            median_phase = medfilt(horizon_phase, kernel_size)
-        # median_phase[full_matrix == Horizon.FILL_VALUE] = 0
+            data_n = data
 
-        img = np.minimum(median_phase - horizon_phase, 2 * np.pi + horizon_phase - median_phase)
-        # img[full_matrix == Horizon.FILL_VALUE] = 0
-        img = np.where(img < -np.pi, img + 2 * np. pi, img)
+        # Generate support coordinates
+        if isinstance(supports, int):
+            if safe_strip:
+                bad_traces_ = bad_traces.copy()
+                bad_traces_[:, :safe_strip], bad_traces_[:, -safe_strip:] = 1, 1
+                bad_traces_[:safe_strip, :], bad_traces_[-safe_strip:, :] = 1, 1
+            else:
+                bad_traces_ = bad_traces
 
-        metric = np.zeros((*img.shape, 2+self.data.shape[2]))
-        metric[:, :, 0] = img
-        metric[:, :, 1] = median_phase
-        metric[:, :, 2:] = phase
+            valid_traces = xp.where(bad_traces_ == 0)
+            indices = xp.random.choice(len(valid_traces[0]), supports)
+            support_coords = xp.asarray([valid_traces[0][indices], valid_traces[1][indices]]).T
 
-        title = 'phase by {}'.format(correction)
+        elif isinstance(supports, (tuple, list, np.ndarray)):
+            support_coords = xp.asarray(supports)
+
+        # Save for plot and introspection
+        self._last_evaluation['support_coords'] = from_device(support_coords)
+
+        # Generate support traces
+        support_traces = data_n[support_coords[:, 0], support_coords[:, 1]]
+        support_stds = data_stds[support_coords[:, 0], support_coords[:, 1]]
+
+        # Compute metric
+        pbar = Notifier(pbar, total=len(support_traces)) if pbar else None
+        accumulator = Accumulator(agg=agg, amortize=amortize, axis=axis, total=len(support_traces))
+        for i, _ in enumerate(support_traces):
+            computed = function(data_n, support_traces[i], data_stds, support_stds[i])
+            computed[bad_traces == 1] = xp.nan
+            accumulator.update(computed)
+            if pbar:
+                pbar.update()
+        if pbar:
+            pbar.close()
+
+        result = accumulator.get(final=True)
+        return from_device(result)
+
+
+    def local_corrs(self, kernel_size=3, normalize=True, agg='mean', amortize=False,
+                    device='cpu', pbar=None, **kwargs):
+        """ Compute correlation in a local fashion. """
+        metric = self.compute_local(function=correlation, data=self.data, bad_traces=self.bad_traces,
+                                    kernel_size=kernel_size, normalize=normalize, agg=agg, amortize=amortize,
+                                    device=device, pbar=pbar)
+
+        title, plot_defaults = self.get_plot_defaults()
+        title = f'Local correlation, k={kernel_size}, with `{agg}` aggregation\nfor {title}'
         plot_dict = {
-            'spatial': self.spatial,
-            'title': '{} for {} on cube {}'.format(title, self.name, self.cube_name),
-            'cmap': METRIC_CMAP,
-            'zmin': -1, 'zmax': 1,
-            'xlabel': 'INLINE_3D', 'ylabel': 'CROSSLINE_3D',
+            **plot_defaults,
+            'title': title,
+            'zmin': -1.0, 'zmax': 1.0,
+            **kwargs
+        }
+        return metric, plot_dict
+
+    def support_corrs(self, supports=100, safe_strip=0, normalize=True, agg='mean', amortize=False,
+                      device='cpu', pbar=None, **kwargs):
+        """ Compute correlation against reference traces. """
+        metric = self.compute_support(function=correlation, data=self.data, bad_traces=self.bad_traces,
+                                      supports=supports, safe_strip=safe_strip,
+                                      normalize=normalize, agg=agg, device=device, amortize=amortize,
+                                      pbar=pbar)
+
+        title, plot_defaults = self.get_plot_defaults()
+        n_supports = supports if isinstance(supports, int) else len(supports)
+        title = f'Support correlation with {n_supports} supports with `{agg}` aggregation\nfor {title}'
+        plot_dict = {
+            **plot_defaults,
+            'title': title,
+            'zmin': -1.0, 'zmax': 1.0,
             **kwargs
         }
         return metric, plot_dict
 
 
-    def quality_map(self, quantiles, metric_names=None, computed_metrics=None, reduce_func='nanmean',
+    def local_crosscorrs(self, kernel_size=3, normalize=False, agg='mean', amortize=False,
+                         device='cpu', pbar=None, **kwargs):
+        """ Compute cross-correlation in a local fashion. """
+        metric = self.compute_local(function=crosscorrelation, data=self.data, bad_traces=self.bad_traces,
+                                    kernel_size=kernel_size, normalize=normalize, agg=agg, amortize=amortize,
+                                    device=device, pbar=pbar)
+        zvalue = np.nanquantile(np.abs(metric), 0.98).astype(np.int32)
+
+        title, plot_defaults = self.get_plot_defaults()
+        title = f'Local cross-correlation, k={kernel_size}, with `{agg}` aggregation\nfor {title}'
+        plot_dict = {
+            **plot_defaults,
+            'title': title,
+            'cmap': 'seismic_r',
+            'zmin': -zvalue, 'zmax': zvalue,
+            **kwargs
+        }
+        return metric, plot_dict
+
+    def support_crosscorrs(self, supports=100, safe_strip=0, normalize=False, agg='mean', amortize=False,
+                           device='cpu', pbar=None, **kwargs):
+        """ Compute cross-correlation against reference traces. """
+        metric = self.compute_support(function=crosscorrelation, data=self.data, bad_traces=self.bad_traces,
+                                      supports=supports, safe_strip=safe_strip,
+                                      normalize=normalize, agg=agg, amortize=amortize, device=device, pbar=pbar)
+        zvalue = np.nanquantile(np.abs(metric), 0.98).astype(np.int32)
+
+        title, plot_defaults = self.get_plot_defaults()
+        n_supports = supports if isinstance(supports, int) else len(supports)
+        title = f'Support cross-correlation with {n_supports} supports with `{agg}` aggregation\nfor {title}'
+        plot_dict = {
+            **plot_defaults,
+            'title': title,
+            'cmap': 'seismic_r',
+            'zmin': -zvalue, 'zmax': zvalue,
+            **kwargs
+        }
+        return metric, plot_dict
+
+
+    def local_btch(self, kernel_size=3, normalize=False, agg='mean', amortize=False,
+                   device='cpu', pbar=None, **kwargs):
+        """ Compute Bhattacharyya divergence in a local fashion. """
+        metric = self.compute_local(function=btch, data=self.probs, bad_traces=self.bad_traces,
+                                    kernel_size=kernel_size, normalize=normalize, agg=agg, amortize=amortize,
+                                    device=device, pbar=pbar)
+
+        title, plot_defaults = self.get_plot_defaults()
+        title = f'Local Bhattacharyya divergence, k={kernel_size}, with `{agg}` aggregation\nfor {title}'
+        plot_dict = {
+            **plot_defaults,
+            'title': title,
+            'zmin': 0.0, 'zmax': 1.0,
+            **kwargs
+        }
+        return metric, plot_dict
+
+    def support_btch(self, supports=100, safe_strip=0, normalize=False, agg='mean', amortize=False,
+                     device='cpu', pbar=None, **kwargs):
+        """ Compute Bhattacharyya divergence against reference traces. """
+        metric = self.compute_support(function=btch, data=self.probs, bad_traces=self.bad_traces,
+                                      supports=supports, safe_strip=safe_strip,
+                                      normalize=normalize, agg=agg, amortize=amortize, device=device, pbar=pbar)
+
+        title, plot_defaults = self.get_plot_defaults()
+        n_supports = supports if isinstance(supports, int) else len(supports)
+        title = f'Support Bhattacharyya divergence with {n_supports} supports with `{agg}` aggregation\nfor {title}'
+        plot_dict = {
+            **plot_defaults,
+            'title': title,
+            'zmin': 0.0, 'zmax': 1.0,
+            **kwargs
+        }
+        return metric, plot_dict
+
+
+    def local_kl(self, kernel_size=3, normalize=False, agg='mean', amortize=False,
+                 device='cpu', pbar=None, **kwargs):
+        """ Compute Kullback-Leibler divergence in a local fashion. """
+        metric = self.compute_local(function=kl, data=self.probs, bad_traces=self.bad_traces,
+                                    kernel_size=kernel_size, normalize=normalize, agg=agg, amortize=amortize,
+                                    device=device, pbar=pbar)
+
+        title, plot_defaults = self.get_plot_defaults()
+        title = f'Local KL divergence, k={kernel_size}, with `{agg}` aggregation\nfor {title}'
+        plot_dict = {
+            **plot_defaults,
+            'title': title,
+            'zmin': None, 'zmax': None,
+            **kwargs
+        }
+        return metric, plot_dict
+
+    def support_kl(self, supports=100, safe_strip=0, normalize=False, agg='mean', amortize=False,
+                   device='cpu', pbar=None, **kwargs):
+        """ Compute Kullback-Leibler divergence against reference traces. """
+        metric = self.compute_support(function=kl, data=self.probs, bad_traces=self.bad_traces,
+                                      supports=supports, safe_strip=safe_strip,
+                                      normalize=normalize, agg=agg, amortize=amortize,
+                                      device=device, pbar=pbar)
+
+        title, plot_defaults = self.get_plot_defaults()
+        n_supports = supports if isinstance(supports, int) else len(supports)
+        title = f'Support KL divergence with {n_supports} supports with `{agg}` aggregation\nfor {title}'
+        plot_dict = {
+            **plot_defaults,
+            'title': title,
+            'zmin': None, 'zmax': None,
+            **kwargs
+        }
+        return metric, plot_dict
+
+
+    def local_js(self, kernel_size=3, normalize=False, agg='mean', amortize=False, device='cpu', pbar=None, **kwargs):
+        """ Compute Jensen-Shannon divergence in a local fashion. """
+        metric = self.compute_local(function=js, data=self.probs, bad_traces=self.bad_traces,
+                                    kernel_size=kernel_size, normalize=normalize, agg=agg, amortize=amortize,
+                                    device=device, pbar=pbar)
+
+        title, plot_defaults = self.get_plot_defaults()
+        title = f'Local JS divergence, k={kernel_size}, with `{agg}` aggregation\nfor {title}'
+        plot_dict = {
+            **plot_defaults,
+            'title': title,
+            'zmin': None, 'zmax': None,
+            **kwargs
+        }
+        return metric, plot_dict
+
+    def support_js(self, supports=100, safe_strip=0, normalize=False, agg='mean', amortize=False,
+                   device='cpu', pbar=None, **kwargs):
+        """ Compute Jensen-Shannon divergence against reference traces. """
+        metric = self.compute_support(function=js, data=self.probs, bad_traces=self.bad_traces,
+                                      supports=supports, safe_strip=safe_strip,
+                                      normalize=normalize, agg=agg, amortize=amortize,
+                                      device=device, pbar=pbar)
+
+        title, plot_defaults = self.get_plot_defaults()
+        n_supports = supports if isinstance(supports, int) else len(supports)
+        title = f'Support JS divergence with {n_supports} supports with `{agg}` aggregation\nfor {title}'
+        plot_dict = {
+            **plot_defaults,
+            'title': title,
+            'zmin': None, 'zmax': None,
+            **kwargs
+        }
+        return metric, plot_dict
+
+
+    def local_hellinger(self, kernel_size=3, normalize=False, agg='mean', amortize=False,
+                        device='cpu', pbar=None, **kwargs):
+        """ Compute Hellinger distance in a local fashion. """
+        metric = self.compute_local(function=hellinger, data=self.probs, bad_traces=self.bad_traces,
+                                    kernel_size=kernel_size, normalize=normalize, agg=agg, amortize=amortize,
+                                    device=device, pbar=pbar)
+
+        title, plot_defaults = self.get_plot_defaults()
+        title = f'Local Hellinger distance, k={kernel_size}, with `{agg}` aggregation\nfor {title}'
+        plot_dict = {
+            **plot_defaults,
+            'title': title,
+            'zmin': None, 'zmax': None,
+            **kwargs
+        }
+        return metric, plot_dict
+
+    def support_hellinger(self, supports=100, safe_strip=0, normalize=False, agg='mean', amortize=False,
+                          device='cpu', pbar=None, **kwargs):
+        """ Compute Hellinger distance against reference traces. """
+        metric = self.compute_support(function=hellinger, data=self.probs, bad_traces=self.bad_traces,
+                                      supports=supports, safe_strip=safe_strip,
+                                      normalize=normalize, agg=agg, amortize=amortize,
+                                      device=device, pbar=pbar)
+
+        title, plot_defaults = self.get_plot_defaults()
+        n_supports = supports if isinstance(supports, int) else len(supports)
+        title = f'Support Hellinger distance with {n_supports} supports with `{agg}` aggregation\nfor {title}'
+        plot_dict = {
+            **plot_defaults,
+            'title': title,
+            'zmin': None, 'zmax': None,
+            **kwargs
+        }
+        return metric, plot_dict
+
+
+    def local_tv(self, kernel_size=3, normalize=False, agg='mean', amortize=False, device='cpu', pbar=None, **kwargs):
+        """ Compute total variation in a local fashion. """
+        metric = self.compute_local(function=tv, data=self.probs, bad_traces=self.bad_traces,
+                                    kernel_size=kernel_size, normalize=normalize, agg=agg, amortize=amortize,
+                                    device=device, pbar=pbar)
+
+        title, plot_defaults = self.get_plot_defaults()
+        title = f'Local total variation, k={kernel_size}, with `{agg}` aggregation\nfor {title}'
+        plot_dict = {
+            **plot_defaults,
+            'title': title,
+            'zmin': None, 'zmax': None,
+            **kwargs
+        }
+        return metric, plot_dict
+
+    def support_tv(self, supports=100, safe_strip=0, normalize=False, agg='mean', amortize=False,
+                   device='cpu', pbar=None, **kwargs):
+        """ Compute total variation against reference traces. """
+        metric = self.compute_support(function=tv, data=self.probs, bad_traces=self.bad_traces,
+                                      supports=supports, safe_strip=safe_strip,
+                                      normalize=normalize, agg=agg, amortize=amortize,
+                                      device=device, pbar=pbar)
+
+        title, plot_defaults = self.get_plot_defaults()
+        n_supports = supports if isinstance(supports, int) else len(supports)
+        title = f'Support total variation with {n_supports} supports with `{agg}` aggregation\nfor {title}'
+        plot_dict = {
+            **plot_defaults,
+            'title': title,
+            'zmin': None, 'zmax': None,
+            **kwargs
+        }
+        return metric, plot_dict
+
+
+    def quality_map(self, quantiles, metric_names=None, computed_metrics=None,
+                    agg='mean', amortize=False, axis=0, apply_smoothing=False,
                     smoothing_params=None, local_params=None, support_params=None, **kwargs):
         """ Create a quality map based on number of metrics.
 
@@ -490,7 +562,7 @@ class BaseSeismicMetric(Metrics):
         reduce_func : str
             Function to reduce multiple metrics into one spatial map.
         smoothing_params, local_params, support_params : dicts
-            Additional parameters for smoothening, local_ metrics, support_ metrics.
+            Additional parameters for smoothening, local metrics, support metrics.
         """
         _ = kwargs
         computed_metrics = computed_metrics or []
@@ -498,38 +570,37 @@ class BaseSeismicMetric(Metrics):
         local_params = local_params or self.LOCAL_DEFAULTS
         support_params = support_params or self.SUPPORT_DEFAULTS
 
-        smoothing_params = {**self.SMOOTHING_DEFAULTS, **smoothing_params}
-        local_params = {**self.LOCAL_DEFAULTS, **local_params}
-        support_params = {**self.SUPPORT_DEFAULTS, **support_params}
+        smoothing_params = {**self.SMOOTHING_DEFAULTS, **kwargs, **smoothing_params}
+        local_params = {**self.LOCAL_DEFAULTS, **kwargs, **local_params}
+        support_params = {**self.SUPPORT_DEFAULTS, **kwargs, **support_params}
 
         if metric_names:
             for metric_name in metric_names:
-                if metric_name.startswith('local'):
+                if 'local' in metric_name:
                     kwds = copy(local_params)
-                elif metric_name.startswith('supp'):
+                elif 'support' in metric_name:
                     kwds = copy(support_params)
 
                 metric = self.evaluate(metric_name, plot=False, **kwds)
                 computed_metrics.append(metric)
 
-        digitized_metrics = []
+        accumulator = Accumulator(agg=agg, amortize=amortize, axis=axis)
         for metric_matrix in computed_metrics:
-            smoothed = smooth_out(metric_matrix, **smoothing_params)
-            digitized = digitize(smoothed, quantiles)
-            digitized_metrics.append(digitized)
+            if apply_smoothing:
+                metric_matrix = smooth_out(metric_matrix, **smoothing_params)
+            digitized = digitize(metric_matrix, quantiles)
+            accumulator.update(digitized)
+        quality_map = accumulator.get(final=True)
 
-        quality_map = np.stack(digitized_metrics, axis=-1)
-        quality_map = getattr(np, reduce_func)(quality_map, axis=-1)
-        quality_map = smooth_out(quality_map, **smoothing_params)
+        if apply_smoothing:
+            quality_map = smooth_out(quality_map, **smoothing_params)
 
-        title = 'quality map'
+        title, plot_defaults = self.get_plot_defaults()
         plot_dict = {
-            'spatial': self.spatial,
-            'title': f'{title} for {self.name} on cube {self.cube_name}',
+            **plot_defaults,
+            'title': f'Quality map for {title}',
             'cmap': 'Reds',
             'zmin': 0.0, 'zmax': np.nanmax(quality_map),
-            'ignore_value': np.nan,
-            'xlabel': 'INLINE_3D', 'ylabel': 'CROSSLINE_3D',
             **kwargs
         }
         return quality_map, plot_dict
@@ -557,11 +628,11 @@ class BaseSeismicMetric(Metrics):
 
 
 
-class HorizonMetrics(BaseSeismicMetric):
+class HorizonMetrics(BaseMetrics):
     """ Evaluate metric(s) on horizon(s).
     During initialization, data along the horizon is cut with the desired parameters.
-    To get the value of a particular metric, use :meth:`.evaluate`:
-    >>> HorizonMetrics(horizon).evaluate('support_corrs', supports=20, agg='mean')
+    To get the value of a particular metric, use :meth:`.evaluate`::
+        HorizonMetrics(horizon).evaluate('support_corrs', supports=20, agg='mean')
 
     To plot the results, set `plot` argument of :meth:`.evaluate` to True.
 
@@ -573,7 +644,7 @@ class HorizonMetrics(BaseSeismicMetric):
         or nested sequence of horizon and list of horizons, then the first horizon is compared against the
         best match from the list.
     other parameters
-        Passed directly to :meth:`.Horizon.get_cube_values` or :meth:`.Horizon.get_cube_values_line`.
+        Passed direcly to :meth:`.Horizon.get_cube_values` or :meth:`.Horizon.get_cube_values_line`.
     """
     AVAILABLE_METRICS = [
         'local_corrs', 'support_corrs',
@@ -581,46 +652,44 @@ class HorizonMetrics(BaseSeismicMetric):
         'local_kl', 'support_kl',
         'local_js', 'support_js',
         'local_hellinger', 'support_hellinger',
-        'local_wasserstein', 'support_wasserstein',
         'local_tv', 'support_tv',
-        'hilbert', 'instantaneous_phase',
+        'instantaneous_phase',
     ]
 
-    def __init__(self, horizons, orientation=None, window=23, offset=0, normalize=False, chunk_size=256, line=1):
+    def __init__(self, horizons, window=23, offset=0, normalize=False, chunk_size=256):
         super().__init__()
         horizons = list(horizons) if isinstance(horizons, tuple) else horizons
         horizons = horizons if isinstance(horizons, list) else [horizons]
         self.horizons = horizons
 
         # Save parameters for later evaluation
-        self.orientation, self.line = orientation, line
         self.window, self.offset, self.normalize, self.chunk_size = window, offset, normalize, chunk_size
 
         # The first horizon is used to evaluate metrics
         self.horizon = horizons[0]
         self.name = self.horizon.name
-        self.cube_name = self.horizon.cube_name
+        self.cube_name = self.horizon.cube_name.replace('amplitudes_', '')
 
-        if orientation is None: # metrics are computed on full cube (spatially)
-            self._data = None # evaluated later
-            self._probs = None
-            self.bad_traces = np.copy(self.horizon.geometry.zero_traces)
-            self.bad_traces[self.horizon.full_matrix == Horizon.FILL_VALUE] = 1
-            self.spatial = True
+        # Properties
+        self._data = None
+        self._probs = None
+        self._bad_traces = None
 
-        else: # metrics are computed on a specific slide
-            self._data, self.bad_traces = self.horizon.get_cube_values_line(orientation=orientation, line=line,
-                                                                            window=window, offset=offset,
-                                                                            normalize=normalize)
-            self._probs = None
-            self.spatial = False
+
+    def get_plot_defaults(self):
+        """ Axis labels and horizon/cube names in the title. """
+        title = f'{self.name} on cube {self.cube_name}'
+        return title, {
+            'xlabel': self.horizon.geometry.axis_names[0],
+            'ylabel': self.horizon.geometry.axis_names[1],
+        }
 
     @property
     def data(self):
         """ Create `data` attribute at the first time of evaluation. """
         if self._data is None:
             self._data = self.horizon.get_cube_values(window=self.window, offset=self.offset,
-                                                      transform=self.normalize, chunk_size=self.chunk_size)
+                                                      normalize=self.normalize, chunk_size=self.chunk_size)
         self._data[self._data == Horizon.FILL_VALUE] = np.nan
         return self._data
 
@@ -628,46 +697,126 @@ class HorizonMetrics(BaseSeismicMetric):
     def probs(self):
         """ Probabilistic interpretation of `data`. """
         if self._probs is None:
-            # Somewhat viable?
-            # mins = np.min(self.data, axis=-1, keepdims=True)
-            # maxs = np.max(self.data, axis=-1, keepdims=True)
-            # shift_scaled = (self.data - mins) / (maxs - mins)
-            # self._probs = shift_scaled / np.sum(shift_scaled, axis=-1, keepdims=True) + self.EPS
-
-            hist_matrix = NumbaNumpy.histo_reduce(self.data, self.horizon.geometry.bins)
+            hist_matrix = histo_reduce(self.data, self.horizon.geometry.bins)
             self._probs = hist_matrix / np.sum(hist_matrix, axis=-1, keepdims=True) + self.EPS
         return self._probs
 
-    def instantaneous_phase(self, **kwargs):
-        """ Compute instantaneous phase via Hilbert transform. """
-        #pylint: disable=unexpected-keyword-arg
-        analytic = hilbert(self.data, axis=2)
+    @property
+    def bad_traces(self):
+        """ Traces to fill with `nan` values. """
+        if self._bad_traces is None:
+            self._bad_traces = self.horizon.geometry.zero_traces.copy()
+            self._bad_traces[self.horizon.full_matrix == Horizon.FILL_VALUE] = 1
+        return self._bad_traces
 
-        phase = np.angle(analytic)
-        phase = phase % (2 * np.pi) - np.pi
 
-        phase_slice = phase[:, :, phase.shape[-1] // 2]
-        phase_slice[np.isnan(np.std(self.data, axis=-1))] = np.nan
+    def perturbed(self, n=5, scale=2.0, clip=3, window=None, kernel_size=3, agg='nanmean', device='cpu', **kwargs):
+        """ Evaluate horizon by:
+            - compute the `local_corrs` metric
+            - perturb the horizon `n` times by random shifts, generated from normal
+            distribution of `scale` std and clipping of `clip` size
+            - compute the `local_corrs` metric for each of the perturbed horizons
+            - get a mean and max value of those metrics: they correspond to the `averagely shifted` and
+            `best generated shifts` horizons
+            - use difference between horizon metric and mean/max metrics of perturbed as a final assesment maps
 
-        avg = mode_scipy(phase_slice[~np.isnan(phase_slice)].round(2), None)
-        phase_slice -= avg[0][0]
-        phase_slice[phase_slice >= np.pi] -= 2 * np.pi
+        Parameters
+        ----------
+        n : int
+            Number of perturbed horizons to generate.
+        scale : number
+            Standard deviation (spread or “width”) of the distribution. Must be non-negative.
+        clip : number
+            Maximum size of allowed shifts
+        window : int or None
+            Size of the data along the height axis to evaluate perturbed horizons.
+            Note that due to shifts, it must be smaller than the original data by atleast 2 * `clip` units.
+        kernel_size, agg, device
+            Parameters of individual metric evaluation
+        """
+        w = self.data.shape[2]
+        window = window or w - 2 * clip
 
-        if np.nanmin(phase_slice) < -np.pi:
-            phase_slice[phase_slice < -np.pi] += 2 * np.pi
-        if np.nanmax(phase_slice) > np.pi:
-            phase_slice[phase_slice > np.pi] -= 2 * np.pi
+        # Compute metrics for multiple perturbed horizons: generate shifts, apply them to data,
+        # evaluate metric on the produced array
+        acc_mean = Accumulator('nanmean')
+        acc_max = Accumulator('nanmax')
+        for _ in range(n):
+            shifts = np.random.normal(scale=2., size=self.data.shape[:2])
+            shifts = np.rint(shifts).astype(np.int32)
+            shifts = np.clip(shifts, -clip, clip)
 
+            shifts[self.horizon.full_matrix == self.horizon.FILL_VALUE] = 0
+            pb = perturb(self.data, shifts, window)
+
+            pb_metric = self.compute_local(function=correlation, data=pb, bad_traces=self.bad_traces,
+                                           kernel_size=kernel_size, normalize=True, agg=agg, device=device)
+            acc_mean.update(pb_metric)
+            acc_max.update(pb_metric)
+
+        pb_mean = acc_mean.get(final=True)
+        pb_max = acc_max.get(final=True)
+
+        # Subtract mean/max maps from the horizon metric
+        horizon_metric = self.compute_local(function=correlation, data=self.data, bad_traces=self.bad_traces,
+                                            kernel_size=kernel_size, normalize=True, agg=agg, device=device)
+        diff_mean = horizon_metric - pb_mean
+        diff_max = horizon_metric - pb_max
+
+        title, plot_defaults = self.get_plot_defaults()
+        title = f'Perturbed metrics\nfor {title}'
         plot_dict = {
-            'spatial': self.spatial,
-            'title': 'Instantaneous phase for {} on cube {}'.format(self.name, self.cube_name),
-            'cmap': 'seismic',
-            'zmin': -np.pi, 'zmax': np.pi,
-            'xlabel': 'INLINE_3D', 'ylabel': 'CROSSLINE_3D',
-            'fill_color': 'black',
+            **plot_defaults,
+            'mode': 'separate',
+            't': title,
+            'titles': ['mean', 'max'],
+            'cmap': 'Reds_r',
+            'zmin': [0.0, -0.5], 'zmax': 0.5,
             **kwargs
         }
-        return phase_slice, plot_dict
+        return (diff_mean, diff_max), plot_dict
+
+
+    def instantaneous_phase(self, device='cpu', **kwargs):
+        """ Compute instantaneous phase via Hilbert transform. """
+        #pylint: disable=unexpected-keyword-arg
+        # Transfer to GPU, if needed
+        data = to_device(self.data, device)
+        xp = cp.get_array_module(data) if (cp is not np) else np
+
+        # Compute hilbert transform and scale to 2pi range
+        analytic = hilbert(data, axis=2)
+
+        phase = xp.angle(analytic)
+        phase = phase % (2 * xp.pi) - xp.pi
+
+        phase_slice = phase[:, :, phase.shape[-1] // 2]
+        phase_slice[np.isnan(xp.std(data, axis=-1))] = xp.nan
+
+        # Evaluate mode value
+        values = phase_slice[~np.isnan(phase_slice)].round(2)
+        _, counts = xp.unique(values, return_counts=True)
+        mode = values[xp.argmax(counts)]
+
+        shifted_slice = phase_slice - mode
+        shifted_slice[shifted_slice >= xp.pi] -= 2 * xp.pi
+
+        # Re-norm so that mode value is at zero point
+        if xp.nanmin(shifted_slice) < -xp.pi:
+            shifted_slice[shifted_slice < -xp.pi] += 2 * xp.pi
+        if xp.nanmax(shifted_slice) > xp.pi:
+            shifted_slice[shifted_slice > xp.pi] -= 2 * xp.pi
+
+        title, plot_defaults = self.get_plot_defaults()
+        title = f'Instantaneous phase\nfor {title}'
+        plot_dict = {
+            **plot_defaults,
+            'title': title,
+            'cmap': 'seismic',
+            'zmin': -np.pi, 'zmax': np.pi,
+            **kwargs
+        }
+        return from_device(shifted_slice), plot_dict
 
 
     def find_best_match(self, offset=0, **kwargs):
@@ -719,27 +868,26 @@ class HorizonMetrics(BaseSeismicMetric):
 
         if printer is not None:
             msg = f"""
-            Comparing horizons:       {self.horizon.name}
-                                      {other.name}
+            Comparing horizons:
+            {self.horizon.name.rjust(45)}
+            {other.name.rjust(45)}
             {'—'*45}
-
-            Rate in 5ms:                         {oinfo['window_rate']:8.4}
-            Mean/std of errors:       {oinfo['mean']:8.4} / {oinfo['std']:8.4}
-            Mean/std of abs errors:   {oinfo['abs_mean']:8.4} / {oinfo['abs_std']:8.4}
-            Max error/abd error:      {oinfo['max']:8} / {oinfo['abs_max']:8}
+            Rate in 5ms:                         {oinfo['window_rate']:8.3f}
+            Mean/std of errors:               {oinfo['mean']:4.2f} / {oinfo['std']:4.2f}
+            Mean/std of abs errors:           {oinfo['abs_mean']:4.2f} / {oinfo['abs_std']:4.2f}
+            Max error/abs error:              {oinfo['max']:4} / {oinfo['abs_max']:4}
             {'—'*45}
-
             Lengths of horizons:                 {len(self.horizon):8}
                                                  {len(other):8}
             {'—'*45}
-            Average heights of horizons:         {(offset + self.horizon.h_mean):8}
-                                                 {other.h_mean:8}
+            Average heights of horizons:         {(offset + self.horizon.h_mean):8.2f}
+                                                 {other.h_mean:8.2f}
             {'—'*45}
-            Coverage of horizons:                {self.horizon.coverage:8.4}
-                                                 {other.coverage:8.4}
+            Coverage of horizons:                {self.horizon.coverage:8.4f}
+                                                 {other.coverage:8.4f}
             {'—'*45}
-            Solidity of horizons:                {self.horizon.solidity:8.4}
-                                                 {other.solidity:8.4}
+            Solidity of horizons:                {self.horizon.solidity:8.4f}
+                                                 {other.solidity:8.4f}
             {'—'*45}
             Number of holes in horizons:         {self.horizon.number_of_holes:8}
                                                  {other.number_of_holes:8}
@@ -775,15 +923,14 @@ class HorizonMetrics(BaseSeismicMetric):
 
 
 
-class GeometryMetrics(BaseSeismicMetric):
-    """ Metrics of cube quality. """
+class GeometryMetrics(BaseMetrics):
+    """ Metrics to asses cube quality. """
     AVAILABLE_METRICS = [
         'local_corrs', 'support_corrs',
         'local_btch', 'support_btch',
         'local_kl', 'support_kl',
         'local_js', 'support_js',
         'local_hellinger', 'support_hellinger',
-        'local_wasserstein', 'support_wasserstein',
         'local_tv', 'support_tv',
     ]
 
@@ -800,9 +947,16 @@ class GeometryMetrics(BaseSeismicMetric):
         self._probs = None
         self._bad_traces = None
 
-        self.spatial = True
         self.name = 'hist_matrix'
         self.cube_name = self.geometry.name
+
+    def get_plot_defaults(self):
+        """ Axis labels and horizon/cube names in the title. """
+        title = f'{self.name} on cube {self.cube_name}'
+        return title, {
+            'xlabel': self.geometry.axis_names[0],
+            'ylabel': self.geometry.axis_names[1],
+        }
 
     @property
     def data(self):
@@ -816,6 +970,7 @@ class GeometryMetrics(BaseSeismicMetric):
         """ Traces to exclude from metric evaluations: bad traces are marked with `1`s. """
         if self._bad_traces is None:
             self._bad_traces = self.geometry.zero_traces
+            self._bad_traces[self.data.max(axis=-1) == self.data.sum(axis=-1)] = 1
         return self._bad_traces
 
     @property
@@ -901,15 +1056,15 @@ class GeometryMetrics(BaseSeismicMetric):
         pbar = tqdm if pbar else lambda iterator, *args, **kwargs: iterator
         metric = np.full((*self.geometries[0].ranges, l), np.nan)
 
-        heights = slice(0, self.geometries[0].cube_shape[2]) if heights is None else slice(*heights)
+        heights = np.arange(self.geometries[0].cube_shape[2]) if heights is None else np.arange(*heights)
 
         with pbar(total=total) as prog_bar:
             for il_block in np.arange(0, self.geometries[0].cube_shape[0], block_size[0]-window[0]):
                 for xl_block in np.arange(0, self.geometries[0].cube_shape[1], block_size[1]-window[1]):
                     block_len = np.min((np.array(self.geometries[0].ranges) - (il_block, xl_block),
                                         block_size), axis=0)
-                    locations = [slice(il_block, il_block + block_len[0]),
-                                 slice(xl_block, xl_block + block_len[1]),
+                    locations = [np.arange(il_block, il_block + block_len[0]),
+                                 np.arange(xl_block, xl_block + block_len[1]),
                                  heights]
 
                     blocks = [prep_func(geometry.load_crop(locations)) for geometry in self.geometries]
@@ -934,679 +1089,3 @@ class GeometryMetrics(BaseSeismicMetric):
             **kwargs
         }
         return metric, plot_dict
-
-
-
-
-# Jit-accelerated NumPy funcions
-@njit
-def geomean(array):
-    """ Geometric mean of an array. """
-    n = np.sum(~np.isnan(array))
-    return np.power(np.nanprod(array), (1 / n))
-
-@njit
-def harmean(array):
-    """ Harmonic mean of an array. """
-    n = np.sum(~np.isnan(array))
-    return n / np.nansum(1 / array)
-
-@njit
-def histo_reduce(data, bins):
-    """ Convert each entry in data to histograms according to `bins`. """
-    i_range, x_range = data.shape[:2]
-
-    hist_matrix = np.full((i_range, x_range, len(bins) - 1), np.nan)
-    for il in prange(i_range):
-        for xl in prange(x_range):
-            hist_matrix[il, xl] = np.histogram(data[il, xl], bins=bins)[0]
-    return hist_matrix
-
-
-class NumbaNumpy:
-    """ Holder for jit-accelerated functions.
-    Note: don't try to automate this with fancy decorators over the function names.
-    """
-    #pylint: disable = unnecessary-lambda, undefined-variable
-    nanmin = njit()(lambda array: np.nanmin(array))
-    nanmax = njit()(lambda array: np.nanmax(array))
-    nanmean = njit()(lambda array: np.nanmean(array))
-    nanstd = njit()(lambda array: np.nanstd(array))
-
-    min = nanmin
-    max = nanmax
-    mean = nanmean
-    std = nanstd
-
-    geomean = geomean
-    harmean = harmean
-
-    histo_reduce = lambda data, bins: histo_reduce(data, bins)
-
-
-
-# Functions to compute metric from data-array
-def compute_local_func(function, name, data, bad_traces, kernel_size=3, reduce_func='nanmean', **kwargs):
-    """ Apply `function` in a `local` way: each entry in `data` is compared against its
-    neighbours, and then those values are reduced to one number with `reduce_func`.
-
-    Parameters
-    ----------
-    function : callable
-        Njitted function to compare two entries in data.
-    name : str
-        Name of function to display on graphs.
-    data : ndarray
-        3D array to apply function to.
-    bad_traces : ndarray
-        Traces to ignore during metric evaluation.
-    kernel_size : int
-        Size of window to reduce values in.
-    reduce_func : str or callable
-        Function to reduce values in window with, e.g. `mean` or `nanmax`.
-    """
-    _ = kwargs
-
-    reduce_func = getattr(NumbaNumpy, reduce_func)
-
-    bad_traces = np.copy(bad_traces)
-    bad_traces[np.std(data, axis=-1) == 0.0] = 1
-
-    padded = np.pad(data, ((kernel_size, kernel_size), (kernel_size, kernel_size), (0, 0)), constant_values=np.nan)
-    bad_traces = np.pad(bad_traces, kernel_size, constant_values=1.0)
-
-    metric = apply_local_func(function, reduce_func, padded, bad_traces, kernel_size)
-    metric = metric[kernel_size:-kernel_size, kernel_size:-kernel_size]
-    title = f'local {name}'
-    return metric, title
-
-
-@njit(parallel=True)
-def apply_local_func(compute_func, reduce_func, data, bad_traces, kernel_size):
-    """ Apply function in window. """
-    #pylint: disable=too-many-nested-blocks, consider-using-enumerate
-    k = int(np.floor(kernel_size / 2))
-    i_range, x_range = data.shape[:2]
-    metric = np.full((i_range, x_range), np.nan)
-
-    for il in prange(i_range):
-        for xl in range(x_range):
-            if bad_traces[il, xl] == 0:
-                trace = data[il, xl, :]
-
-                metric_element = np.full((kernel_size, kernel_size), np.nan)
-                for _idx in range(-k, k+1):
-                    for _jdx in range(-k, k+1):
-                        if bad_traces[il+_idx, xl+_jdx] == 0:
-                            trace_ = data[il+_idx, xl+_jdx]
-                            metric_element[k+_idx, k+_jdx] = compute_func(trace, trace_)
-                metric_element[k, k] = np.nan
-
-                if np.sum(~np.isnan(metric_element)):
-                    metric[il, xl] = reduce_func(metric_element)
-    return metric
-
-
-def compute_support_func(function_ndarray, function_str, name,
-                         data, supports, bad_traces, safe_strip=0, line_no=None, **kwargs):
-    """ Apply function to compare each trace and a number of support traces.
-
-    Parameters
-    ----------
-    supports : int, sequence, ndarray or str
-        Defines mode of generating support traces.
-        If int, then that number of random non-zero traces positions are generated.
-        If sequence or ndarray, then must be of shape (N, 2) and is used as positions of support traces.
-        If str, then must define either `iline` or `xline` mode. In each respective one, iline/xline given by
-        `line_no` argument is used to generate supports.
-    safe_strip : int
-        Used only for `int` mode of `supports` parameter and defines minimum distance
-        from borders for sampled points.
-    line_no : int
-        Used only for `str` mode of `supports` parameter to define exact iline/xline to use.
-
-    Returns
-    -------
-    array-like
-        Matrix of either (n_ilines, n_xlines, n_supports) or (n_ilines, n_xlines) shape with
-        computed metric for each point.
-    """
-    _ = kwargs
-    bad_traces = np.copy(bad_traces)
-    bad_traces[np.std(data, axis=-1) == 0.0] = 1
-
-    if isinstance(supports, (int, tuple, list, np.ndarray)):
-        if isinstance(supports, int):
-            title = f'{name} with {supports} random supports'
-            if safe_strip:
-                bad_traces[:, :safe_strip], bad_traces[:, -safe_strip:] = 1, 1
-                bad_traces[:safe_strip, :], bad_traces[-safe_strip:, :] = 1, 1
-
-            np.random.seed(0)
-            non_zero_traces = np.where(bad_traces == 0)
-            indices = np.random.choice(len(non_zero_traces[0]), supports)
-            supports = np.array([non_zero_traces[0][indices], non_zero_traces[1][indices]]).T
-
-        elif isinstance(supports, (tuple, list, np.ndarray)):
-            title = f'{name} with {len(supports)} supports'
-            if min(len(item) == 2 for item in supports) is False:
-                raise ValueError('Each of `supports` sequence must contain coordinate of trace (il, xl). ')
-            supports = np.array(supports)
-
-        metric = function_ndarray(data, supports, bad_traces)
-
-    elif isinstance(supports, str):
-        if function_str is None:
-            raise ValueError(f'{name} does not work in `line` mode!')
-
-        title = f'{name} on {line_no} {supports}'
-        if supports.startswith('i'):
-            support_il = line_no or data.shape[0] // 2
-            metric = function_str(data, bad_traces, support_il=support_il)
-
-        if supports.startswith('x'):
-            support_xl = line_no or data.shape[1] // 2
-            metric = function_str(data, bad_traces, support_xl=support_xl)
-
-    else:
-        raise ValueError('`Supports` must be either int, sequence, ndarray or string. ')
-    return metric, title
-
-
-
-def compute_local_corrs(data, bad_traces, kernel_size=3, reduce_func='nanmean', **kwargs):
-    """ Compute correlation between each column in data and nearest traces. """
-    return compute_local_func(_compute_local_corrs, 'correlation',
-                              data=data, bad_traces=bad_traces,
-                              kernel_size=kernel_size, reduce_func=reduce_func, **kwargs)
-
-@njit
-def _compute_local_corrs(array_1, array_2):
-    result = np.sum((array_1 - np.mean(array_1)) * (array_2 - np.mean(array_2))) / (np.std(array_1) * np.std(array_2))
-    return result / len(array_1)
-
-
-def compute_support_corrs(data, supports, bad_traces, safe_strip=0, line_no=None, **kwargs):
-    #pylint: disable=missing-function-docstring
-    return compute_support_func(function_ndarray=_compute_support_corrs,
-                                function_str=_compute_line_corrs,
-                                name='correlation',
-                                data=data, supports=supports, bad_traces=bad_traces,
-                                safe_strip=safe_strip, line_no=line_no, **kwargs)
-
-def _compute_support_corrs(data, supports, bad_traces):
-    n_supports = len(supports)
-    i_range, x_range, depth = data.shape
-
-    data_n = data - np.mean(data, axis=-1, keepdims=True)
-    data_stds = np.std(data, axis=-1)
-
-    support_traces = np.zeros((n_supports, depth))
-    for i in range(n_supports):
-        coord = supports[i]
-        support_traces[i, :] = data[coord[0], coord[1], :]
-    support_n = support_traces - np.mean(support_traces, axis=-1, keepdims=True)
-    support_stds = np.std(support_traces, axis=-1)
-
-    corrs = np.zeros((i_range, x_range, n_supports))
-    for i in range(n_supports):
-        cov = np.sum(support_n[i] * data_n, axis=-1) / depth
-        temp = cov / (support_stds[i] * data_stds)
-        temp[bad_traces == 1] = np.nan
-        corrs[:, :, i] = temp
-    return corrs
-
-def _compute_line_corrs(data, bad_traces, support_il=None, support_xl=None):
-    depth = data.shape[-1]
-
-    data_n = data - np.mean(data, axis=-1, keepdims=True)
-    data_stds = np.std(data, axis=-1)
-    bad_traces[data_stds == 0] = 1
-
-    if support_il is not None and support_xl is not None:
-        raise ValueError('Use `compute_support_corrs` for given trace. ')
-
-    if support_il is not None:
-        support_traces = data[[support_il], :, :]
-        support_n = support_traces - np.mean(support_traces, axis=-1, keepdims=True)
-        support_stds = np.std(support_traces, axis=-1)
-        bad_traces[:, support_stds[0, :] == 0] = 1
-    if support_xl is not None:
-        support_traces = data[:, [support_xl], :]
-        support_n = support_traces - np.mean(support_traces, axis=-1, keepdims=True)
-        support_stds = np.std(support_traces, axis=-1)
-        bad_traces[support_stds[:, 0] == 0, :] = 1
-
-    cov = np.sum(support_n * data_n, axis=-1) / depth
-    corrs = cov / (support_stds * data_stds)
-    corrs[bad_traces == 1] = 0
-    return corrs
-
-
-def compute_local_crosscorrs(data, bad_traces, kernel_size=3, reduce_func='nanmean', **kwargs):
-    """ Compute cross-correlation between each column in data and nearest traces. """
-    return compute_local_func(_compute_local_crosscorrs, 'Cross-correlation',
-                              data=data, bad_traces=bad_traces,
-                              kernel_size=kernel_size, reduce_func=reduce_func, **kwargs)
-
-@njit
-def _compute_local_crosscorrs(array_1, array_2):
-    temp = np.zeros(len(array_1))
-    for k in range(len(array_1)):
-        temp[k] = np.sum(array_1[k:] * array_2[:len(array_1)-k])
-    return np.argmax(temp) - (temp.size // 2)
-
-
-def compute_support_crosscorrs(data, supports, bad_traces, safe_strip=0, **kwargs):
-    #pylint: disable=missing-function-docstring
-    return compute_support_func(function_ndarray=_compute_support_crosscorrs,
-                                function_str=None,
-                                name='Cross-correlation',
-                                data=data, supports=supports, bad_traces=bad_traces,
-                                safe_strip=safe_strip, **kwargs)
-
-def _compute_support_crosscorrs(data, supports, bad_traces):
-    n_supports = len(supports)
-    i_range, x_range, depth = data.shape
-
-    support_traces = np.zeros((n_supports, depth*2))
-    for i in range(n_supports):
-        coord = supports[i]
-        support_traces[i, :] = np.pad(data[coord[0], coord[1], :], pad_width=(depth//2, depth - depth//2))
-
-    divs = np.zeros((i_range, x_range, n_supports))
-    for i in range(n_supports):
-        supports_ = support_traces[i]
-        temp = np.zeros((i_range, x_range, depth))
-        for k in range(depth):
-            temp[:, :, k] = np.sum(supports_[k:k+depth] * data, axis=-1)
-        temp = np.argmax(temp, axis=-1).astype(float) - depth//2
-        temp[bad_traces == 1] = np.nan
-        divs[:, :, i] = temp
-    return divs
-
-
-
-def compute_local_btch(data, bad_traces, kernel_size=3, reduce_func='nanmean', **kwargs):
-    """ Compute Bhattacharyya distance between each column in data and nearest traces. """
-    return compute_local_func(_compute_local_btch, 'Bhattacharyya-divergence',
-                              data=data, bad_traces=bad_traces,
-                              kernel_size=kernel_size, reduce_func=reduce_func, **kwargs)
-
-@njit
-def _compute_local_btch(array_1, array_2):
-    return np.sum(np.sqrt(array_1 * array_2))
-
-
-def compute_support_btch(data, supports, bad_traces, safe_strip=0, **kwargs):
-    #pylint: disable=missing-function-docstring
-    return compute_support_func(function_ndarray=_compute_support_btch,
-                                function_str=None,
-                                name='Bhattacharyya-divergence',
-                                data=data, supports=supports, bad_traces=bad_traces,
-                                safe_strip=safe_strip, **kwargs)
-
-def _compute_support_btch(data, supports, bad_traces):
-    n_supports = len(supports)
-    i_range, x_range, depth = data.shape
-
-    support_traces = np.zeros((n_supports, depth))
-    for i in range(n_supports):
-        coord = supports[i]
-        support_traces[i, :] = data[coord[0], coord[1], :]
-
-    divs = np.zeros((i_range, x_range, n_supports))
-    for i in range(n_supports):
-        supports_ = support_traces[i]
-        temp = np.sum(np.sqrt(supports_ * data), axis=-1)
-        temp[bad_traces == 1] = np.nan
-        divs[:, :, i] = temp
-    return divs
-
-
-def compute_local_kl(data, bad_traces, kernel_size=3, reduce_func='nanmean', **kwargs):
-    """ Compute Kullback-Leibler divergence between each column in data and nearest traces. """
-    return compute_local_func(_compute_local_kl, 'KL-divergence',
-                              data=data, bad_traces=bad_traces,
-                              kernel_size=kernel_size, reduce_func=reduce_func, **kwargs)
-
-@njit
-def _compute_local_kl(array_1, array_2):
-    return 1 - np.sum(array_1 * np.log2(array_1 / array_2))
-
-
-def compute_support_kl(data, supports, bad_traces, safe_strip=0, **kwargs):
-    #pylint: disable=missing-function-docstring
-    return compute_support_func(function_ndarray=_compute_support_kl,
-                                function_str=None,
-                                name='KL-divergence',
-                                data=data, supports=supports, bad_traces=bad_traces,
-                                safe_strip=safe_strip, **kwargs)
-
-def _compute_support_kl(data, supports, bad_traces):
-    n_supports = len(supports)
-    i_range, x_range, depth = data.shape
-
-    support_traces = np.zeros((n_supports, depth))
-    for i in range(n_supports):
-        coord = supports[i]
-        support_traces[i, :] = data[coord[0], coord[1], :]
-
-    divs = np.zeros((i_range, x_range, n_supports))
-    for i in range(n_supports):
-        supports_ = support_traces[i]
-        temp = 1 - np.sum(supports_ * np.log2(supports_/data), axis=-1)
-        temp[bad_traces == 1] = np.nan
-        divs[:, :, i] = temp
-    return divs
-
-
-
-def compute_local_js(data, bad_traces, kernel_size=3, reduce_func='nanmean', **kwargs):
-    """ Compute Jensen-Shannon distance between each column in data and nearest traces. """
-    return compute_local_func(_compute_local_js, 'JS-divergence',
-                              data=data, bad_traces=bad_traces,
-                              kernel_size=kernel_size, reduce_func=reduce_func, **kwargs)
-
-@njit
-def _compute_local_js(array_1, array_2):
-    average = (array_1 + array_2) / 2
-    log_average = np.log2(average)
-    div_1 = np.sum(array_1 * (np.log2(array_1) - log_average))
-    div_2 = np.sum(array_2 * (np.log2(array_2) - log_average))
-    return 1 - (div_1 + div_2) / 2
-
-
-def compute_support_js(data, supports, bad_traces, safe_strip=0, **kwargs):
-    #pylint: disable=missing-function-docstring
-    return compute_support_func(function_ndarray=_compute_support_js,
-                                function_str=None,
-                                name='JS-divergence',
-                                data=data, supports=supports, bad_traces=bad_traces,
-                                safe_strip=safe_strip, **kwargs)
-
-def _compute_support_js(data, supports, bad_traces):
-    n_supports = len(supports)
-    i_range, x_range, depth = data.shape
-
-    support_traces = np.zeros((n_supports, depth))
-    for i in range(n_supports):
-        coord = supports[i]
-        support_traces[i, :] = data[coord[0], coord[1], :]
-
-    divs = np.zeros((i_range, x_range, n_supports))
-    for i in range(n_supports):
-        supports_ = support_traces[i]
-
-        average = (supports_ + data) / 2
-        log_average = np.log2(average)
-        div_1 = np.sum(supports_ * (np.log2(supports_) - log_average), axis=-1)
-        div_2 = np.sum(data * (np.log2(data) - log_average), axis=-1)
-        temp = 1 - (div_1 + div_2) / 2
-        temp[bad_traces == 1] = np.nan
-        divs[:, :, i] = temp
-    return divs
-
-
-
-def compute_local_hellinger(data, bad_traces, kernel_size=3, reduce_func='nanmean', **kwargs):
-    """ Compute Hellinger distance between each column in data and nearest traces. """
-    return compute_local_func(_compute_local_hellinger, 'hellinger distance',
-                              data=data, bad_traces=bad_traces,
-                              kernel_size=kernel_size, reduce_func=reduce_func, **kwargs)
-
-SQRT_2 = np.sqrt(2)
-@njit
-def _compute_local_hellinger(array_1, array_2):
-    return 1 - np.sqrt(np.sum(np.sqrt(array_1) - np.sqrt(array_2)) ** 2) / SQRT_2
-
-
-def compute_support_hellinger(data, supports, bad_traces, safe_strip=0, **kwargs):
-    #pylint: disable=missing-function-docstring
-    return compute_support_func(function_ndarray=_compute_support_hellinger,
-                                function_str=None,
-                                name='hellinger distance',
-                                data=data, supports=supports, bad_traces=bad_traces,
-                                safe_strip=safe_strip, **kwargs)
-
-def _compute_support_hellinger(data, supports, bad_traces):
-    n_supports = len(supports)
-    i_range, x_range, depth = data.shape
-
-    support_traces = np.zeros((n_supports, depth))
-    for i in range(n_supports):
-        coord = supports[i]
-        support_traces[i, :] = data[coord[0], coord[1], :]
-
-    dist = np.zeros((i_range, x_range, n_supports))
-    for i in range(n_supports):
-        temp = 1 - np.sqrt(np.sum((np.sqrt(support_traces[i]) - np.sqrt(data)) ** 2, axis=-1)) / SQRT_2
-        temp[bad_traces == 1] = np.nan
-        dist[:, :, i] = temp
-    return dist
-
-
-
-def compute_local_wasserstein(data, bad_traces, kernel_size=3, reduce_func='nanmean', **kwargs):
-    """ Compute Wasserstein distance between each column in data and nearest traces. """
-    return compute_local_func(_compute_local_wasserstein, 'wesserstein distance',
-                              data=data, bad_traces=bad_traces,
-                              kernel_size=kernel_size, reduce_func=reduce_func, **kwargs)
-
-@njit
-def _compute_local_wasserstein(array_1, array_2):
-    sorter_1 = np.argsort(array_1)
-    sorter_2 = np.argsort(array_2)
-
-    concatted = np.concatenate((array_1, array_2))
-    concatted = np.sort(concatted)
-    deltas = np.diff(concatted)
-
-    cdf_indices_1 = np.searchsorted(array_1[sorter_1], concatted[:-1], 'right')
-    cdf_indices_2 = np.searchsorted(array_2[sorter_2], concatted[:-1], 'right')
-
-    cdf_1 = cdf_indices_1 / array_1.size
-    cdf_2 = cdf_indices_2 / array_2.size
-    return 1 - np.sum(np.multiply(np.abs(cdf_1 - cdf_2), deltas))
-
-
-def compute_support_wasserstein(data, supports, bad_traces, safe_strip=0, **kwargs):
-    #pylint: disable=missing-function-docstring
-    return compute_support_func(function_ndarray=_compute_support_wasserstein,
-                                function_str=None,
-                                name='wasserstein distance',
-                                data=data, supports=supports, bad_traces=bad_traces,
-                                safe_strip=safe_strip, **kwargs)
-
-def _compute_support_wasserstein(data, supports, bad_traces):
-    n_supports = len(supports)
-    i_range, x_range, depth = data.shape
-
-    support_traces = np.zeros((n_supports, depth))
-    for i in range(n_supports):
-        coord = supports[i]
-        support_traces[i, :] = data[coord[0], coord[1], :]
-
-    divs = np.zeros((i_range, x_range, n_supports))
-    for i in range(n_supports):
-        temp = _emd_array(support_traces[i], data)
-        temp[bad_traces == 1] = np.nan
-        divs[:, :, i] = temp
-    return divs
-
-@njit
-def _emd_array(array_1d, array_3d):
-    temp = np.zeros(array_3d.shape[:2])
-
-    for i in prange(array_3d.shape[0]):
-        for j in range(array_3d.shape[1]):
-            temp[i, j] = _compute_local_wasserstein(array_1d, array_3d[i, j, :])
-    return temp
-
-
-
-def compute_local_tv(data, bad_traces, kernel_size=3, reduce_func='nanmean', **kwargs):
-    """ Compute Bhattacharyya distance between each column in data and nearest traces. """
-    return compute_local_func(_compute_local_tv, 'Total variation',
-                              data=data, bad_traces=bad_traces,
-                              kernel_size=kernel_size, reduce_func=reduce_func, **kwargs)
-
-@njit
-def _compute_local_tv(array_1, array_2):
-    return 1 - 0.5*np.sum(np.abs(array_1 - array_2))
-
-
-def compute_support_tv(data, supports, bad_traces, safe_strip=0, **kwargs):
-    #pylint: disable=missing-function-docstring
-    return compute_support_func(function_ndarray=_compute_support_tv,
-                                function_str=None,
-                                name='Total variation',
-                                data=data, supports=supports, bad_traces=bad_traces,
-                                safe_strip=safe_strip, **kwargs)
-
-def _compute_support_tv(data, supports, bad_traces):
-    n_supports = len(supports)
-    i_range, x_range, depth = data.shape
-
-    support_traces = np.zeros((n_supports, depth))
-    for i in range(n_supports):
-        coord = supports[i]
-        support_traces[i, :] = data[coord[0], coord[1], :]
-
-    divs = np.zeros((i_range, x_range, n_supports))
-    for i in range(n_supports):
-        supports_ = support_traces[i]
-        temp = 1 - 0.5*np.sum(np.abs(supports_ - data), axis=-1)
-        temp[bad_traces == 1] = np.nan
-        divs[:, :, i] = temp
-    return divs
-
-
-
-def smooth_out(matrix, kernel_size=3, sigma=2.0, iters=3, **kwargs):
-    """ Convolve the matrix with gaussian kernel with special treatment to `np.nan`s:
-    if the point is not `np.nan`, then it is changed to a weighted sum of all present points nearby.
-
-    Parameters
-    ----------
-    kernel_size : int
-        Size of gaussian filter.
-    sigma : number
-        Standard deviation (spread or “width”) for gaussian kernel.
-        The lower, the more weight is put into the point itself.
-    """
-    _ = kwargs
-    k = int(np.floor(kernel_size / 2))
-
-    ax = np.linspace(-(kernel_size - 1) / 2., (kernel_size - 1) / 2., kernel_size)
-    x_points, y_points = np.meshgrid(ax, ax)
-    kernel = np.exp(-0.5 * (np.square(x_points) + np.square(y_points)) / np.square(sigma))
-    gaussian_kernel = (kernel / np.sum(kernel).astype(np.float32))
-
-    smoothed = np.copy(matrix)
-    smoothed = np.pad(smoothed, kernel_size, constant_values=np.nan)
-
-    for _ in range(iters):
-        smoothed = apply_local_smoothing(smoothed, k, gaussian_kernel.ravel())
-    smoothed = smoothed[kernel_size:-kernel_size, kernel_size:-kernel_size]
-    smoothed[np.isnan(matrix)] = np.nan
-    return smoothed
-
-@njit
-def apply_local_smoothing(matrix, k, raveled_kernel):
-    """ Jit-accelerated function to apply smoothing. """
-    #pylint: disable=too-many-nested-blocks, consider-using-enumerate
-    i_range, x_range = matrix.shape
-    smoothed = np.full((i_range, x_range), np.nan)
-
-    for iline in prange(i_range):
-        for xline in prange(x_range):
-
-            if not np.isnan(matrix[iline, xline]):
-                element = matrix[iline-k:iline+k+1, xline-k:xline+k+1]
-
-                s, sum_weights = 0.0, 0.0
-                for item, weight in zip(element.ravel(), raveled_kernel):
-                    if not np.isnan(item):
-                        s += item * weight
-                        sum_weights += weight
-
-                if sum_weights != 0.0:
-                    val = s / sum_weights
-                    smoothed[iline, xline] = val
-
-    return smoothed
-
-
-def digitize(matrix, quantiles):
-    """ Convert continious metric into binarized version with thresholds defined by `quantiles`. """
-    bins = np.nanquantile(matrix.ravel(), np.sort(quantiles)[::-1])
-
-    if len(bins) > 1:
-        digitized = np.digitize(matrix, [*bins, np.nan]).astype(float)
-        digitized[digitized > 0] -= 1
-    else:
-        digitized = np.zeros_like(matrix, dtype=np.float64)
-        digitized[matrix <= bins[0]] = 1.0
-
-    digitized[np.isnan(matrix)] = np.nan
-    return digitized
-
-
-def gridify(matrix, frequencies, iline=True, xline=True):
-    """ Convert digitized map into grid with various frequencies corresponding to different bins. """
-    values = np.unique(matrix[~np.isnan(matrix)])
-    if len(values) != len(frequencies):
-        min_freq = min(frequencies)
-        max_freq = max(frequencies)
-        multiplier = np.power(max_freq/min_freq, 1/(len(values) - 1))
-        frequencies = [np.rint(max_freq / (multiplier ** i))
-                       for i, _ in enumerate(values)]
-    else:
-        frequencies = np.sort(frequencies)[::-1]
-
-    grid = np.zeros_like(matrix)
-    for value, freq in zip(values, frequencies):
-        idx_1, idx_2 = np.nonzero(matrix == value)
-
-        if iline:
-            mask = (idx_1 % freq == 0)
-            grid[idx_1[mask], idx_2[mask]] = 1
-        if xline:
-            mask = (idx_2 % freq == 0)
-            grid[idx_1[mask], idx_2[mask]] = 1
-
-    grid[np.isnan(matrix)] = np.nan
-    return grid
-
-
-
-def enlarge_carcass_metric(metric, geometry, width=10):
-    """ Increase visibility of a sparce metric grid. """
-    structure = np.ones((1, 3), dtype=np.uint8)
-    metric = np.copy(metric)
-    metric[np.isnan(metric)] = Horizon.FILL_VALUE
-    dilated_1 = cv2.dilate(metric, structure, iterations=width)
-    dilated_2 = cv2.dilate(metric, structure.T, iterations=width)
-
-    metric = np.full_like(metric, np.nan)
-    metric[dilated_1 > -999] = dilated_1[dilated_1 > -999]
-    metric[dilated_2 > -999] = dilated_2[dilated_2 > -999]
-    metric[(dilated_1 > -999) & (dilated_2 > -999)] = (dilated_1[[(dilated_1 > -999) & (dilated_2 > -999)]] + \
-                                                       dilated_2[[(dilated_1 > -999) & (dilated_2 > -999)]]) / 2
-
-    metric[np.isnan(geometry.std_matrix)] = np.nan
-    return metric
-
-
-
-@njit
-def correct_pi(horizon_phase, eps):
-    """ Jit-accelerated function to <>. """
-    for i in range(horizon_phase.shape[0]):
-        prev = horizon_phase[i, 0]
-        for j in range(1, horizon_phase.shape[1] - 1):
-            if np.abs(np.abs(prev) - np.pi) <= eps and np.abs(np.abs(horizon_phase[i, j + 1]) - np.pi) <= eps:
-                horizon_phase[i, j] = prev
-            prev = horizon_phase[i, j]
-    return horizon_phase
